Fix memory leak
If the cursor goes to the left, the tape gets allocated again, copied into the new tape and the new tape replaces the old one.
But the old tape array was never freed, which means that this is memory which is on the heap, whilst not having a referencing pointer.